agent/requirements.txt
livekit
livekit-api
livekit-agents[silero]
livekit-plugins-openai
livekit-plugins-silero
python-dotenv>=0.19.0
requests>=2.26.0

# Database - PostgreSQL with asyncpg for high-concurrency
asyncpg>=0.29.0
psycopg2-binary>=2.9.0  # For connection utilities

# RAG - Document indexing and vector search
sentence-transformers>=2.2.0  # For embeddings
aiofiles>=23.0.0  # Async file operations
docling>=2.0.0  # Advanced document parsing (PDF, DOCX, PPTX, HTML, tables, OCR)

# LLM Providers - HTTP clients with connection pooling
aiohttp>=3.9.0  # Async HTTP client for LLM providers
httpx>=0.25.0  # Alternative async HTTP client (used for Ollama embeddings)
vllm  # Local LLM inference server

# Enhanced features (optional - uncomment to enable)
# LangChain for conversation memory and RAG
# langchain>=0.1.0
# langchain-community>=0.1.0
# chromadb>=0.4.0
# sentence-transformers>=2.2.0

# Premium STT/TTS providers (require API keys)
# livekit-plugins-deepgram    # Better noise handling
# livekit-plugins-elevenlabs  # High-quality TTS
# livekit-plugins-google      # Google Speech
# livekit-plugins-azure       # Azure Speech
# livekit-plugins-cartesia    # Fast TTS
# livekit-plugins-assemblyai  # Real-time STT
# livekit-plugins-anthropic   # Claude LLM